+++ packages/javascript/lib/format/index.ts
@@ -1,5 +1,7 @@
 export { duration } from './duration.js';
+export { durationParts } from './durationParts.js';
 export { fileSizeFormat } from './fileSizeFormat.js';
+export { fileSizeParts } from './fileSizeParts.js';
 export { numberFormat } from './numberFormat.js';
 export { safeJSONParse } from './safeJSONParse.js';
 export { safeParseInt } from './safeParseInt.js';